IDEA-313553 Fix 'show hidden tabs' button behaviour in other non-editor places where tabs are used
Generalize the logic of showing icon and close action to not depend on "Editor tabs" settings (that is intended to be only for editor).
Hardcode painting of the selected tab underline on the left side, because default TabPainter support only bottom position for underline.
Also, specify to show scrollbar only in the EditorTabs. The scrollbar should not be shown by default in all inheritors of the JBEditorTabs.

(cherry picked from commit f0267cfd558fd8e299fe81878c5b4f40f5dddcbc)

IJ-CR-103204

GitOrigin-RevId: dcdf8587478970144c077de2eb77b8e7a7a9a6b5