[UI] IDEA-313422 Move JBColor cache creation to LafManagerImpl
This ensures that JBColor doesn't modify UI Defaults from a BGT.
Also make the cache thread-safe using ConcurrentHashMap,
as it's actually used from different threads.

Since ConcurrentHashMap doesn't allow nulls, we need to use
a special object for the null value.

There are if (cache != null) checks just in case JBColor is ever
used before initializing LaF, but that doesn't seem to be the case.
Anyway, even if it happens, we just won't be caching values
until LaF is initialized.

GitOrigin-RevId: dc8eb6338bd4a98178ec6f18aa818d9cf65c019a