[UI] IDEA-313380 Defer setting weight of tool window if root pane has no size
Sometimes, for example, when a project is being opened, it can happen
so that the root pane has no size yet as it has only been just added.
This lead to the tool window size being set to zero, effectively to the
minimum size.

Defer setting weight in such cases, but take care to cancel the scheduled
weight setting if setWeight() is called again for the same anchor.

GitOrigin-RevId: 26d71b6107e0983c22de2cb67e9ba657e72553dd