Fix LaF Defaults internal action
This action obtain all keys using `UIManager.getDefaults().keys()`, that creates HashSet from entries, so hashcode value of JBColor is required. But there are also stored `JBColor.CACHED_NULL` values, that throw an exception when hashcode method is invoked. So, it is required to specify hashcode function for such marker objects.

GitOrigin-RevId: bc66322a58625bacc1f961581569decaf605bf2f