IDEA-142666 Fix position of the light bulb
If there is enough space at the beginning of the caret line, put the
light bulb there. If that doesn't work, try the line above the caret
line, and fall back to the line below the caret line.

Align the popup menu with the light bulb panel. Ensure that the popup
menu doesn't cover the caret line.

Previously, the light bulb could reach into the gutter area or the
editor tabs, now it is restricted to the editor content component.

Previously, the light bulb was not vertically centered in the caret
line, its exact position depended on the zoom level.

Previously, the popup menu could overlap the light bulb panel by more
than a single pixel.

Previously, if the editor font was small, the light bulb panel could
cover text from the lines above or below the caret line.

Previously, the position calculation involved expressions that were hard
to understand, such as "icon height + 1 border", even though the panel
has a top and a bottom border. The magic number '3' did not have an
obvious meaning, it may have been NORMAL_BORDER_SIZE/2.

Previously, the light bulb was placed directly adjacent to text in the
caret line, now there is a 2-character safety margin, to avoid clicking
the light bulb when selecting the text of the line.

GitOrigin-RevId: cd4e6ce9310e2128992cfe61ac1f7cf52b8e032f