Fix flaky GradleRelativeConfigCalculatingTest
The main idea of this test is to create two identical Gradle projects (even the root folder should be the same), but existing in different neighboring folders. This parent folder should not be part of the project, so we will have identical projects from the root, but they will have different absolute paths. And relativiser must write the same config.dat so that we can re-use it using the portable caches.

GitOrigin-RevId: 5f97c481c3521466b5c94943ddc77d3d35994465