PY-56541 Remove creating a nested context to detect the scope owner of a TypeVar
It's difficult to reason about how these should work with caching of
PyTypingTypeProvider.Context in TypeEvalContext.

Additionally, I fixed losing an already found TypeVar scope for "T" when
constructing a PyGenericType for "Type[T]". Previously, we did it all over again
for Type[T] in getTypeForResolvedElement at the moment where the corresponding
TypeVar declaration already wasn't at the top of the type alias stack, so it
started to cause EmptyStack exceptions after this change.

GitOrigin-RevId: b22f2ca0ccd12ad9a21294d6ba23574d839e813b