IDEA-314175: Handle relative classpath jar paths
We convert paths to absolute paths because JarLoader and
clients thereof are unprepared for relative paths.

This change is necessary for certain environments (such as Bazel tests)
in which relative paths are used for the sake of hermeticity.
Previously, using PathClassLoader in Bazel would result in invalid
resource paths and other runtime errors.

This change should have no effect on existing environments where
absolute paths are already used.