[UI] IDEA-93140 Check for exising names when validating new layout name
It was already done for renaming, but not for creating a new layout.

It's kind of ugly that the validator is now placed in the same file as
the RenameLayoutAction, as it's used in more than one action.
The best way would be to convert remaining actions to Java
and put them all in a single Kotlin file, but it doesn't look like
a good time to do it now, when the release branch is almost created.

GitOrigin-RevId: 95d6ec9630f9806999305e6b946142f6acdc117e