[UI] IDEA-162125 Add registry key to make browse button focusable
It isn't focusable for historical reasons, and after an internal discussion
in Slack it was decided not to break user habits by suddenly making
it focusable. But Google wants it focusable for Android Studio, so
make a key to let them change it in their product.

IJ-CR-103770

(cherry picked from commit d4d4baad77a6eaaf235a3199c18123232e567bb0)

Change-Id: Id4d4baad77a6eaaf235a3199c18123232e567bb0