[UI] IDEA-162125 Add registry key to make browse button focusable
It isn't focusable for historical reasons, and after an internal discussion
in Slack it was decided not to break user habits by suddenly making
it focusable. But Google wants it focusable for Android Studio, so
make a key to let them change it in their product.

(cherry picked from commit 9cc63d0679260c9a2f0ded9729b4dca738715700)

IJ-CR-103770

GitOrigin-RevId: 0a29865b553e4545d95e16e0c8af9fce918e614f