[UI] IDEA-93140 Fix renaming layout to the same name
Old logic: first we add the new layout, then remove the old one.

New logic: vice versa.

The old logic had a bug because if the old and new names are
the same, we first 'add' (not really) the new layout, and then
remove it, thus effectively deleting it.

It should never happen anyway because the rename dialog
doesn't allow renaming to the same name, but it's no excuse
for a bug in the back-end logic.

GitOrigin-RevId: 0f7cd09a088cadcda83c680684f12e7027dda9b6