[UI] IDEA-93140 Validate layout names according to guidelines
The layout name input dialog is extracted into a separate class.
A rather complicated validator is implemented.

It would be best to implement this in the API, but it would
take a lot of effort to design an extensible, flexible and
convenient to use API for all use cases.

To implement validation according to the guidelines,
the dialog and validator need to consider at what
moments validation happens: when the dialog is
first shown (don't show an error message right
away even if the initial value is incorrect, which
it is), when the user types something or when
the OK button is pressed (and to make sure
validation happens at all when it's pressed,
we need to keep it enabled by faking a successful
validation attempt before that).

There's also a rather ugly hack to break the
circle of the dialog and validator initialization.
Due to the way both are designed, the only
way is to store a reference to the validator in
a global variable during the initialization.
Which is, of course, not something one is usually
willing to do, but does the trick.

We could, of course, just make the field protected,
but that would be ugly as well, and we better keep
the ugliness at the implementation side, not in the API.

GitOrigin-RevId: 948db74467ea25c50ac70ee1352987e4332f430e