[IFT] Fix debug toolbar highlighting
Found that there can be 'Debugger' tab instead of 'Threads & Variables', so it is better to not rely on a tab name.
Also found that delay is still required here, because selected tab is changing quite often after opening the debug tool window. The required tab is selected first, then deselected and selected again.

GitOrigin-RevId: c4fbe189859a10e747078e696856eefb633493ed