[UI] IDEA-93140 Layout Name Dialog: don't associate error message with field
It only works when the field is either focused or hovered. If the user types
an invalid name and presses OK with the mouse, the error message doesn't
appear because the OK button is both focused and hovered, not the input field.

It's best fixed in the dialog itself, but for now just use the error message
not connected to a field: it's displayed under the field, and is always there.

(cherry picked from commit 82a3ead6c5e9ebcf3304c639a4defa24f8db263f)
(cherry picked from commit c1523a757037d0d25dbd2a335862142de119bce4)

IJ-CR-104180

GitOrigin-RevId: b658033e174146a260091a6ed09d0e6e431f1790