[UI] IDEA-93140 Layout Name Dialog: don't associate error message with field
It only works when the field is either focused or hovered. If the user types
an invalid name and presses OK with the mouse, the error message doesn't
appear because the OK button is both focused and hovered, not the input field.

It's best fixed in the dialog itself, but for now just use the error message
not connected to a field: it's displayed under the field, and is always there.

GitOrigin-RevId: 82a3ead6c5e9ebcf3304c639a4defa24f8db263f