[workspace model index] add assertion about access to uninitialized index (IDEA-313304)
If some code accesses WorkspaceFileIndex before it's initialized in ModuleBridgeLoaderService, we shouldn't initialize it because otherwise it'll be updated inside write action when the entities are loaded and this may cause freeze. So now we return empty data instead, and report an error because most probably code which access the project configuration before the project is fully loaded is incorrect, such activities should be postponed.

GitOrigin-RevId: 1f9df3f1c2545c7e2d71c9470a82ba5eb68d434c