[kotlin] Test for injected kotlin snippet
This commit introduces IDE side test for KT-56632.
Through the highlighting it makes sure that a configuration is successfully
loaded for an injected Kotlin code snippet.

Usually highlighting tests are launched with explicitly forbidden
`canChangeDocumentDuringHighlighting` whereas IDEA itself is not. The flag
prevents not only from a document update but from PSI and a model ones as well.
Since the new test deals with scripts (injected code is treated as scripts) we face
project model update (`ScriptClassRootsUpdater.applyDiffToModel`).
Having the flag set to "IDEA mode" is ok for injected code
(`InjectedLanguageHighlightingTest#testInjectedFragmentsHighlightedIncrementallyAsSoonAsTheyAvailableStress`)

GitOrigin-RevId: 880c03484597313d346f8af87cbbee89d6ef2500