[UI] IDEA-93140 Make restore layout actions non-toggleable
They were effectively toggleable, but never selected. Unfortunately,
this led to extra indent in menus, reserved for the check mark
that was never there.

This doesn't apply to the quick switch actions, where check marks
are actually needed.

This led to some code duplication, but it's just two lines of code.
Still, it needs to be refactored someday.

GitOrigin-RevId: 61abdd6d5762b2b7a52c3e3b4f22781d47ba7428