[vcs-log] introduce api for more flexible column preferred width calculation
Introduce PreferredWidth interface to represent how column width is calculated and allow VcsLogCellRenderer to select one of the two strategies: use fixed width or calculate the width from data (the latter strategy may be resource consuming, so needs to be separate). The new api allows to move specifics of the calculation to the VcsLogCellRenderer itself instead of hardcoding them in VcsLogGraphTable.

GitOrigin-RevId: 31602dc07308259340c07135b6edce651b7159b6