[UI] IDEA-93140 Fall back to default layout without changing name
When the active layout contains no descriptors, for example, because
it was saved for the old UI and now the new UI is active, or vice versa,
then the loadDefaultLayout() method would be invoked that would reset
the active layout to "Custom" with the default descriptors. This would
repeat again and again, until the next "Save Layout" action.

Fix by falling back to the default descriptors if there aren't any, but
keep the active layout name. This will keep the active layout at all
times, which is obviously the expected behavior. In case the layout
was created for the old UI and then the new UI was activated, it would
now lead to the layout effectively containing the default descriptors
for the new UI, which is the expected behavior.

(cherry picked from commit 17958eff84127ecd8049d74800b5a0c6346754c1)
(cherry picked from commit c85113541c70775c26ddff143e034016d56c8245)

IJ-CR-104180

GitOrigin-RevId: 6fa542310df6ba129b588b2c533f61964c4d71bf