[java] set compiler output path for project even if some importers were found (IDEA-313418)
Even if files from build system are detected under an opened directory, it still makes sense to set some default value for the compiler output path. It may happen that at least some modules won't be imported from the build systems, and we need to have a meaningful default output directory for that in such a case.

GitOrigin-RevId: 3bad73c50b4f47b2ec4e2e5e99e7d2ccb45b1fdc