[UI] IDEA-309823 Temporary fix for cut off hints/mnemonics
There were several issues with how the preferred size of the myMnemonicLabel
was calculated:

- the font was set after the size is calculated;
- the size was calculated using a temporary label, not the one actually used;
- the most important: the size didn't take the border into account,
  precisely because the temporary label was used (that didn't have the border).

For now, fix by temporarily setting the text to the same label that will be used
later. Set the font before calculating the size.

The fix is temporary because it only applies to the new UI and because
instead of just using the "W" string, ideally we should iterate over all possible
mnemonics and take the maximum size. Possibly caching the result to avoid
recomputing it every time (font calculations are usually slow).

GitOrigin-RevId: ef96a3ca8e650ab900403a0c0f829dae3b822770