[inlays] take into account enabled providers in tests
It's only possible to disable a group of providers, not a specific provider. Thus, tests should accept a group ID to be enabled to avoid confusion. Also, the previous implementation wasn't working and actually all providers stayed enabled as one can see by PHP tests. Even though we could continue passing a provider ID and extract a group ID out of it, it might create a wrong impression that a test can be run only for a particular provider.

GitOrigin-RevId: 0b83282ce930438c8e5ec011882f8da88baf8192