[platform] IDEA-313422 Document threading rules for CustomStatusBarWidget
It's yet another of the cases when a component is not really a Swing component,
but more like an API entity that wraps a Swing component. The said entity
can be instantiated on a BGT, a Swing component can not, as it can lead
to CME when Swing decides to modify some internal structures in a constructor,
specifically UI Defaults.

GitOrigin-RevId: d53371d32adedc9938382b2ccf818eb102ea36d8