[platform] IDEA-313422 Document threading rules for StatusBarWidget
It's yet another of the cases when a component is not really a Swing component,
but more like an API entity that wraps a Swing component. The said entity
can be instantiated on a BGT, a Swing component can not, as it can lead
to CME when Swing decides to modify some internal structures in a constructor,
specifically UI Defaults.

GitOrigin-RevId: 1cc9e7823a44c236a740acf37038c3f3b99d35a5