[new terminal] Edit completion algorithm to catch the moment when all completion items are written
When all completion items are written - terminal returns the caret to the prompt line and writes the command again. If there are more lines than screen, it writes the whole prompt with the command. So, we can consider that moments as the end of the completion and parse all items at once.

GitOrigin-RevId: 8c3b62ba8ca3be1e4458e5df88c06396deb1b9ef