Fix parser for breakpoint comments in debugger tests
Allow breakpoint comments at the end of a line, leaving it to the
caller to decide on which line to place the breakpoint.

Allow custom types of breakpoints, not only 'Method' and 'Line'.

Ensure that a breakpoint comment does not have duplicate properties.

In error messages, fix the off-by-one error in the reported line
numbers.

In error messages about invalid boolean values, add the column number.

In Java debugger tests, make the location of parse errors clickable.

GitOrigin-RevId: 555b28e603551995fc4fb0c89a52bf8caa515d4f