[kotlin] fix stepping out of inline lambdas
Previously when trying to step out of an inline lambda, the debugger would always perform a line-wise stepping out. This behaviour resulted in errors when trying to step out of one-line inline lambdas - you would always jump to the next line. This commit sets the minimal stepping size when trying to step out of an inline lambda.

^KTIJ-24239 Fixed

GitOrigin-RevId: d5fc4a6fa8efb6e3b3648e232c7d3a16a4e31d4e