[platform] EA-658585 Remove unnecessary canSelect() call
The only thing SelectInContext.selectIn() does that
SelectInTarget.selectIn() doesn't is check canSelect() first
and return false if the target can't select the desired element.

However, since selectInCurrentTarget() doesn't really check
the return value, that check seems to be unnecessary. We just
need to ensure that SelectInTarget.selectIn() won't break
(and just do nothing) if it's impossible to select the desired element.
This seems to be the case, but if it ever breaks, we just need to
fix that part.

GitOrigin-RevId: a3015e0fd100071a0f2793413b76047c39b247b9