Register WebP support as an ImageIO plugin
Currently, WebP support is manually registered on startup using the
application service mechanism. For this it tries to get the IIORegistry
default instance. If it had not yet been created, but is in the process
of being created in another thread, that will lead to a race condition
where only one of the instances will be retained and the WebP support
might be lost.
Instead of manually registering the WebP support, this makes use of the
plugin facility of ImageIO, where plugins that are present in the
classpath will be automatically registered.
Defining an ImageIO plugin requires having files in META-INF/services
which specify which services are implemented and where to find their
implementations.