[UI] IDEA-313096 Remove ProjectViewImpl.myCurrentSelectionObsolete
It was introduced to fix IDEA-206431, but it doesn't reproduce anymore
even without this hack. Probably due to some changes in the way auto
scrolling is handled when the Project View is shown.

It's necessary to remove this because toolWindowShown() in the
constructor writes to this field before calling selectInCurrentTarget().
However, to get rid of slow ops in the EDT related to the selectInContext,
we need to merge findSelectInContext() and selectInCurrentTarget()
into a single call and get rid of the null check. This is needed because
once we migrate the relevant PSI code into a BGT, it'll be no longer
possible to perform this null check in the EDT, at least not that early,
as we'll need to complete BGT computations first.

Since this is the only place that had any meaningful code between
the null check and the selectInCurrentTarget() call, we can now
proceed to merging the code and then to splitting it between a BGT
and the EDT.

GitOrigin-RevId: 2c4792aaccdc35bbe0f859a056029d6b6f909e6b