[workspace file index] remove registry key which allows users to switch back to the old implementation (IDEA-276394)
It was enabled by default in all JetBrains IDEs (except Rider) in 2022.3 releases, and Rider also enabled this option by default in 2023.1. No major regressions were found, so now we can remove this option. This will allow us to use WorkspaceFileIndex in new code without providing a fallback. Also, we'll be able to remove the old implementation later.

GitOrigin-RevId: 8ca9fea12b81c87302d710946e236bb61ce6a359