[IFT] Use simplified shift key presentation on Mac
Found that users often interpret '⇧' glyph as up arrow button when shortcut consist only of single '⇧' glyph. So, it would be better to write it with a description in this case - '⇧ Shift'.
But when shift is used as modifier for some other key - it will still be represented as single glyph.

(cherry picked from commit 00aa30f7d8f21fb02732bd729c0aec38c9ba52f0)

IJ-CR-104697

GitOrigin-RevId: 98dc5aecaeb18ced9c3605c1044385e8b8c411f6