[project model] support resetting custom kinds for libraries
Before it wasn't possible to remove a custom kind and properties from an existing library, but it's needed for Scala plugin (logic which imports Scala SDK libraries from sbt changes to fix SCL-18866, and it's needed to reset kinds for previously imported libraries). Removing and recreating the library without a kind doesn't work in some cases because of code in ProjectModifiableLibraryTableBridgeImpl.prepareForCommit. So now it's possible to pass 'null' to ModifiableModelEx.setKind, and implementations are updated accordingly.

GitOrigin-RevId: 9ed26a0d9fdc0a96c95f77eef51a57ca29eb3160