[UI] IDEA-264248 Show child modules of a module in Packages View
There are module groups and there are modules that can also act
as module groups. The previous implementation only returned child
modules for genuine module groups, but for modules it only returned
files/directories and libraries. When a module was used to group
submodules, this led to empty modules being shown.

Fix by adding child modules as well, using the ModuleGrouper API.

GitOrigin-RevId: 5d9f67c5b2489c4906ad07927931ad9e4a6b99b9