Fix unstable and hanging DumbServiceImplTest
Instability was caused by presence of two instances of DumbServiceImpl, each of which was publishing DumbModeListener events. In the end shared SmartModeScheduler might go into inconsistent state when project's DumbService is SMART, but scheduler thinks that it is DUMB because of mixed events from two sources.

Hangs were caused by absence of API for timed waiting. Now this API is introduced.

GitOrigin-RevId: b8ddcaa9518854baa9b5eb02a9eb08b8ef0922da