[platform] deprecate ProjectNameProvider.getDefaultName and remove its usage (EA-819352)
This method is called during the project initialization, and if the implementation accesses passed Project instance, it may lead to exceptions (see EA-819352). So it's dangerous to have this method in its current form. Moreover, the method wasn't used for the purpose it was introduced (IDEA-171878), and now 'getNameFile' method can be used for that purpose, so it's better to deprecate the method.

GitOrigin-RevId: b6b0c5dc17e43a5814a8547a966ae82613f81fef