[intellij project configuration] convert source roots in intellij.python.helpers to resource roots
Since there are no Java source files in these directories, it doesn't make sense to register them as source roots. Converting to resource roots should speed up compilation a little, and also makes it simpler to write tests which analyze class files in the module output (we can report an error if there are no class files in a module with source roots).

GitOrigin-RevId: 4efe77ef07c8e2449186319e806abe7835e3e763