[UI] IDEA-315412 Convert Tool Window Layouts action group into DefaultActionGroup
This way its children are not visible in Customize Toolbar, as DefaultActionGroup
gets special treatment in ActionsTreeUtil.getActions(), only the children that are
defined in PlatformActions.xml are visible there.

Override update() to make it a popup action group by default, for example,
when it's added to a toolbar, but not in the main menu, where it's already
in a (popup group) submenu.

(cherry picked from commit dd910ea40a9ca8658151fcd956055456edd3e9c0)

IJ-CR-105005

GitOrigin-RevId: 8c487a9bcf57782d78b875ec26477a2bceb6e885