[UI] IDEA-315412 Convert Tool Window Layouts action group into DefaultActionGroup
This way its children are not visible in Customize Toolbar, as DefaultActionGroup
gets special treatment in ActionsTreeUtil.getActions(), only the children that are
defined in PlatformActions.xml are visible there.

Override update() to make it a popup action group by default, for example,
when it's added to a toolbar, but not in the main menu, where it's already
in a (popup group) submenu.

GitOrigin-RevId: dd910ea40a9ca8658151fcd956055456edd3e9c0