[debugger] Fix smart step into interface methods overridden by delegation in K2 code
The difference between K1 and K2 here is that the latter sets line number
in generated overrides of methods implemented by delegation.
As a result, JDI stops at the line of "by" keyword when we do STEP_INTO.
But we don't understand that we're in delegated method, and we do STEP_OUT.

The K2 behaviour looks more correct, so the fix here is from the debugger side.

Also, an ability to add custom .out for K2 test was added because output
differs expectedly.

See KTIJ-24739

(cherry picked from commit 53106089deab377333cbe24b45a32061e924de42)

IJ-MR-105785

GitOrigin-RevId: 133eab9d874807461f71dde4a35cf5b6e57af6bc