[UI] IDEA-315658 Decrease separator insets in Git Branches popup
It turns out to be more complicated than it would seem.
Since it's a tree popup, it uses Tree.rowHeight for all items, including
separators. So decreasing insets will only shift the separator within
the row, but the row height will remain as is.

To fix it, set rowHeight to 0 for the new UI and then remove the
extra insets. In order for it to work, though, the height for all
non-separator rows is now reported by the renderer as
the preferred size, otherwise rows just have the smallest height
possible.

(cherry picked from commit 9892567e0248daea370ead24f23a77eaf5be428e)

IJ-CR-105249

GitOrigin-RevId: 4d1803ca8931ad0a74bd755c7c338e97fb84ad6b