[UI] IDEA-315998 Fix blinking navbar on UI settings change
It was removed and re-added (via invokeLater, of all things).
Fix by not doing that unless its location actually changed.

While we're at it, fix the previous fix for the same thing in the status bar.
That fix was only applied when it's actually in the status bar, now when
we fixed the top position we can never add/remove it no matter
where it is, unless its location actually changed.

GitOrigin-RevId: 36db65c608255ea847f8e29371fbaf803092b8ef