IDEA-315948: xdebugger: Don't request navigation in updateExecutionPosition()
The initial navigation to the new execution point is performed anyway,
and the extra call makes it also navigate to the old position once that
position is gone. It happens because the navigation request isn't
synchronized with the execution point state and when the new `null`
execution point was set, the navigation request was handled by the
old non-null execution point.

Fixes IDEA-315948 Debugger navigates to the stopping point on resume

GitOrigin-RevId: 2d79155a95c8fb4f098164046ea0980dc25ec6f4