[UI] IDEA-314903 Update global editor color scheme in case of delegation
MyColorSchemeDelegate.updateGlobalScheme would not update the scheme
if it's not the default one (as denoted by myCustomGlobalScheme == null).
However, this custom scheme could actually be a scheme delegating to
the default global scheme, and in this case it must be updated, otherwise
it isn't switched properly when the global scheme is changed, which
is why consoles wouldn't update properly on a theme change.

Fix by going through the delegation chain until either a non-delegating
custom scheme is found (which should not be updated) or a scheme
delegating to the global default is found, which is then updated as usual.

(cherry picked from commit 4da6b79342903f29c70fba0f4f2985e132546ce7)

IJ-CR-105345

GitOrigin-RevId: c312260b788b24a3faddf850296c3221f063c956