[UI] Replace Color with JBColor for Console ANSI colors
To make ANSI colors adapt to the current theme, use
JBColor with a Supplier to get whatever color corresponds
to the ANSI value.

The downside is, of course, the potential performance hit.
I've checked the code and it doesn't look like something
VERY taxing, but of course it's worse than simply returning
some constant value. If it becomes a problem we can
try caching it at some point.

GitOrigin-RevId: 0b02bbc05431c5f05e532ac74b5ecd44a6aa7db1