[kotlin] Fix for exception on resolving enum classes
Fix for exception "Resolver for 'project source roots
and libraries <...> does not know how to resolve".

According to exception analyzer, "infos" is empty in
`DiagnoseUnknownModuleInfoReporter.errorInSdkResolver`.
Which means no modules found in
`ProjectResolutionFacade.resolverForElement`.
We try to use `resolverForModuleInfo` instead, which
doesn't need modules. For this we can use
`getIdeService` instead of `tryGetFrontendService` in
`ResolutionFacade`.

^KTIJ-25083 Fixed

GitOrigin-RevId: 4f1cf8b8cc2e7576d9a525372f428401dec0b9bc