[kotlin] Fix for exception on resolving enum classes
Fix for exception "Resolver for 'project source roots
and libraries <...> does not know how to resolve".

According to exception analyzer, "infos" is empty in
`DiagnoseUnknownModuleInfoReporter.errorInSdkResolver`.
Which means no modules found in
`ProjectResolutionFacade.resolverForElement`.
We try to use `resolverForModuleInfo` instead, which
doesn't need modules. For this we can use
`getIdeService` instead of `tryGetFrontendService` in
`ResolutionFacade`.

^KTIJ-25083 Fixed

GitOrigin-RevId: 5a581880b487bd0398246e9eeb92bd9d95e9c5eb