[UI] IDEA-315962 Fix InfoAndProgressPanel EDT initialization
The previous attempt was not exactly correct because it created
a lot of independently initialized and not properly encapsulated
lazy fields. This, in particular, caused the bug with the navbar
would not be properly removed because the relevant field
(myShowNavBar) was not initialized and set to false even
though the navbar was actually displayed.

Fix by moving most of the EDT code into an inner class,
InfoAndProgressPanelImpl. It needs further cleanup,
as its fields are sometimes accessed in a messy way,
and it also accesses the outer class rather randomly.
But it already fixed the problem with the navbar.

GitOrigin-RevId: 51d8d3f9e92a3b8a2c5f35e926f2f95701c5e101