PY-59727 Infer types from usages only for collections assigned to variables
For everything else, e.g. d['key'] = {'foo': 42}, we cannot track subsequent
modifications of d['key'], such as d['key']['bar'] = 1, anyway. We need
a reference as a subscription operand, or a callee to make sure these mutating
operations are performed on the same value. It doesn't make sense to scan
through an entire containing scope looking for subscription expressions
and calls on such assignments targets.

GitOrigin-RevId: 8ebbcb73e43e5711cbe680da6936f3f1c4e54a55