[KTIJ-25101] Fix an AlreadyDisposedException
The module for a given Kotlin script is not necessarily the same
throughout the lifetime of a project; a sync might destroy and
re-create a module, meaning that we must recompute the module
each time we compute the target platform (rather than re-using
the module we initially computed).

Move the call to getScriptModule to close to its use, making it
clear that one is immediately-used and the other is at an
arbitrarily later time.

closes https://github.com/JetBrains/intellij-community/pull/2367

^KTIJ-25101 fixed

GitOrigin-RevId: f75cad77fd0699b41e90b1e9fd81e5a4cc563432