Fix an AlreadyDisposedException
The module for a given Kotlin script is not necessarily the same
throughout the lifetime of a project; a sync might destroy and
re-create a module, meaning that we must recompute the module
each time we compute the target platform (rather than re-using
the module we initially computed).

Move the call to getScriptModule to close to its use, making it
clear that one is immediately-used and the other is at an
arbitrarily later time.