[UI] IDEA-316461 Fix restoring layout for unavailable tool windows
In case if a tool window is unavailable, there were some bugs due
to trying to create their stripe buttons and even show them if
they were visible when the layout was stored.

Fix by first making all unavailable tool windows invisible in the new
layout regardless of what isVisible value was stored.

Then, apply their info as usual, so it can be used later, if that window
becomes available and is shown due to some user action.

And lastly, don't try to do anything else with them at the time
the layout is restored (simply remove them from the to-do list).
Since they're not available, we can safely assume they are not
visible, so trying to change their state in any way would be
a useless exercise in futility.

GitOrigin-RevId: da04574dde9b65baa6603cbfc1953e59cd4bf81c