[platform] IDEA-315659 Create DefaultToolWindowLayoutExtension
DefaultToolWindowLayoutProvider was problematic as it couldn't be used
by multiple plugins, as only one override could exist for the service.

The new API allows for multiple extensions to be registered. Their
invocation order can be controlled explicitly, and they're passed the
same builder instance, so they can extend, modify and reuse the previous
state, or reset it and start from scratch, up to their choice.

Old-to-new and new-to-old adapters were added. The default tool window
layout manager only uses the new API now, but there's an extension
that invokes the old API and converts its results. The old API implementation,
on the other hand, delegates to the default new API implementation.

Once Rider migrates to the new API, we can safely remove both of these
bridges and use only the new API.

GitOrigin-RevId: 395cdcdf19fc0c99f356ae7c8ae29b003ade68f2