IDEA-317154 Fix wrong Find/Replace button border after updateUI()
Every updateUI() call would reset the button's border to
the default specified in the theme, which directly refers to
the DarculaButtonPainter class. However, since these buttons
actually use a customized subclass with a smaller border,
this would reset the border on every updateUI() call,
for example, a theme switch or Compact Mode toggle.

This didn't lead to any unusual effects without Compact Mode,
but in Compact Mode it leads to those buttons being higher
than they're supposed to be, which in turn would stretch the
Replace field's border, making it a double one.

Fix by subclassing JButton and overriding updateUI()
to explicitly update the border every time.

GitOrigin-RevId: 0ce750b4648aeb5c41b167e0958f52c3ebfd3920