[UI] IDEA-317153 Fix font update for JBTextArea/JBTextField placeholder
The initial idea (IDEA-311878) was to reset the font to the parent's
font on updateUI(). However, this didn't really work because
updateUI() is called in post-order, so the parent's font is actually
updated AFTER the child's updateUI() has finished. It sort-of-worked
for cases when the parent's font isn't really customized, so the only
kind of update that was required was for the font to scale properly,
which it somehow did (didn't look into it, but sounds reasonably
possible). However, when the font was really different from default,
it didn't work as it was only updated in the parent's updateUI().

Fix by resetting the font to null instead of to the parent's font.
This causes the check in paintStatusText() to trigger on the next
paint, effectively resetting the font to the parent's, which at
that time is, of course, up-to-date already.

(cherry picked from commit cbb4312febd2ce61127a3261307152435fae53e3)

IJ-CR-105870

GitOrigin-RevId: 94505b3f8518dbdd357ea1090944591347b22143