[UI] IDEA-316669 Don't paint background image over MainToolbar
Whether it was painted or not was determined by the point in the Swing
hierarchy at which painting was started. If it involved a JBPanel, then
it would create a graphics delegate that would then be passed to all children.

Fix by explicitly unwrapping that graphics in MainToolbar.getComponentGraphics.
Note that the fix will only work as long as there's just a single graphics delegate
and not, for example, a whole chain of them. This is the case currently, though.

GitOrigin-RevId: c93f12c7466f14601a6d84ed9340444d89467ad1