[project model] refactoring: move some code from ModuleManagerComponentBridge to projectModel.impl module (IDEA-302252)
ModuleManagerComponentBridge is located in intellij.platform.lang.impl module, but most of its code doesn't depend on classes of this module and may be moved to intellij.platform.projectModel.impl module where the rest of the project model implementation is located. This is needed to simplify the hierarchy of classes and avoid problems in codeServer modules which don't depend on intellij.platform.lang.impl but may require this functionality.

GitOrigin-RevId: c687430f7a514a74feccbfecd8a16d7f511acc9e