[devkit] IDEA-317283 Use correct resolve scope in IncorrectParentDisposableInspection
The old version of the code used `resolveScope` of the resolved `psiMethod`
to look for `Project`,  `Application` and `Module` classes. However,
the `psiMethod` might be declared in the module which have no dependency
on those classes - for example, a method like `Disposer.register`.

This change uses the file call site scope to make sure that the classes
in question are found. If they are not, it means that they cannot be used
on the call site either, so no need for the inspection

N.B.: IncorrectParentDisposableInspectionTest didn't catch that problem because
in the test environment all dependencies see each other, so the scope of
the `Disposer.register` could see the required classes

^IDEA-317283 Fixed

GitOrigin-RevId: d56eacc841b69c093ea3af5cdabe8db0fb2366fe