[platform] initial version of runtime module repository (IJPL-109)
This change introduces RuntimeModuleRepository API which will be used to determine paths to classes and other resources of modules at runtime. It also adds RuntimeModuleRepositoryBuilder which generates descriptors for modules during compilation of intellij project (this is controlled by a registry key which is switched off by default for now).

GitOrigin-RevId: be33235ec0cdb7ea55a925e753b9b58a3ca63ac3