PY-60054 - clear "pydev_message" and always set result of evaluated breakpoint expression
The previous implementation of "pydevd_frame.py" updated the "pydev_message" only if the evaluated breakpoint expression did not return "None". In some cases, this can result in stale messages from a previously evaluated breakpoint expression.

The "ProtocolParser" now no longer discards the message if it is an empty string or "None". Both are valid values for evaluated breakpoint expressions. Therefore, the "pydev_message" is now initialized with None instead of an empty string.